1.16.2 Compat?
Probally wont work since I dont really know how to code, Im doing a shot in the dark that will probally miss, but hey! Goal is to fix issue #5 . Which Im under the impression is based on the Custom Biome Code of 1.16.2. As the issue started the same snapshot that this feature appeared.